+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/pm/SshSoftFencingCommand.java
@@ -121,7 +121,7 @@ public class SshSoftFencingCommand<T extends VdsActionParameters> extends VdsCom
      */
     private boolean checkIfHostBecomeUp() {
         VdsManager vdsManager = getResourceManager().getVdsManager(getVdsId());
-        int sleepInterval = Config.<Integer> getValue(ConfigValues.VdsRefreshRate) * 1000;
+        long sleepInterval = Config.<Integer> getValue(ConfigValues.VdsRefreshRate) * 1000;
         while (vdsManager.isHostInGracePeriod(true)) {
             if (vdsManager.getCopyVds().getStatus() == VDSStatus.Up) {
                 // host became Up during grace period
